Create `challenges` folder

+++ challenges/Set_1/eight.py
@@ -1,8 +1,7 @@
-from Crypto.Cipher import AES
 import binascii
 
 def get_16_bytes_blocks_frequency(ciphertext):
-    blocks = [ciphertext[i*16:(i+1)*16] for i in range(len(ciphertext) / 16)]
+    blocks = [ciphertext[i*16:(i+1)*16] for i in range(int(len(ciphertext) / 16))]
     return len(set(blocks))
 
 if __name__ == '__main__':
@@ -18,4 +17,4 @@ if __name__ == '__main__':
             minScore = curScore
 
     print("The best candidate is {}... with {} repeating blocks".format(winner[:16],
-                                                                     len(winner)/16 - minScore))
+                                                                     int(len(winner)/16) - minScore))
